+++ CHANGES.rst
@@ -5,7 +5,9 @@ Changelog
 1.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Support changing PDF watermark image on publication.
+  `#2385 <https://github.com/syslabcom/scrum/issues/2385>`_
+  [reinhardt]
 
 
 1.0.0 (2024-06-12)